Fix OSGetTick.
It's a u32 not an s64.

+++ include/revolution/os.h
@@ -68,7 +68,7 @@ bool OSIsEnableInterrupts();
 /*
 * Real-time clock API
 */
-s64 OSGetTick();
+u32 OSGetTick();
 u32 OSGetTime();
 
 #ifdef __cplusplus
